\name{p_minus}
\alias{p_minus}
\title{
p_minus
}
\description{
This is the function to compute p-, which is required for the Mann-Whitney-Wilcoxon
test.
}
\usage{
p_minus(p1, p2)
}
\arguments{
  \item{p1}{
A vector containing the probabilities of the outcome falling into each level of the
control arm.
}
  \item{p2}{
A vector containing the probabilities of the outcome falling into each level of the
experimental arm.
}
}
\value{
The value of p-.
}
\keyword{internal}
