\name{OptOTrials-package}
\alias{OptOTrials-package}
\alias{OptOTrials}
\docType{package}
\title{
\packageTitle{OptOTrials}
}
\description{
\packageDescription{OptOTrials}
}
\details{
There are several main functions. \code{Decision_rule_S_1stage},
\code{Decision_rule_M_1stage}, \code{Decision_rule_W_1stage}, \code{ruleF}, and
\code{ruleFS} determine the decision rule for clinical trials. \code{op.1stage},
\code{op.F}, and \code{op.FS} calculate the operating characteristics for clinical trial
designs, including type I error, power, and expected sample size, to investigate the
performance of the designs.
}
\author{
\packageAuthor{OptOTrials}

Maintainer: \packageMaintainer{OptOTrials}
}
\references{
Park, Y. (2025). Optimal two-stage group sequential designs based on
Mann-Whitney-Wilcoxon test. PloS one, 20(2), e0318211.
}
%\keyword{Adaptive design, Mann-Whitney-Wilcoxon test, score test, sample size, win odds test}

%% Uncomment below to imitate parts of library(help = OptOTrials)
%\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{OptOTrials}}
%\section{Documentation Index}{\packageIndices{OptOTrials}}


