\name{p_plus}
\alias{p_plus}
\title{
p_plus
}
\description{
This is the function to comupute p+, which is required for the Mann-Whitney-Wilcoxon test. 
}
\usage{
p_plus(p1, p2)
}
\arguments{
  \item{p1}{
A vector containing the probabilities of the outcome falling into each level of the control arm.
}
  \item{p2}{
A vector containging the probabilities of the outcome falling into each level of the control arm.
}
}
\value{
The value of p+.
}
