\name{theta}
\alias{theta}
\title{
Calculation of theta
}
\description{
This is the function to compute theta (i.e., the expectation of T), which is required for the Win Odds test. 
}
\usage{
theta(p1, p2)
}
\arguments{
  \item{p1}{
A vector containing the probabilities of the outcome falling into each level of the control arm.
}
  \item{p2}{
A vector containging the probabilities of the outcome falling into each level of the control arm.
}
}
\value{
The value of theta
}
